old-backend/requirements.txt
boto3==1.24.20
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.11.0
django-filter==23.2
django-imagekit==4.1.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-phonenumber-field==6.3.0
django-silk==5.0.3
django-storages==1.12.3
django-timezone-field==5.0
drf-yasg==1.20.0
exponent-server-sdk==2.0.0
holidays==0.14.2
icalendar==5.0.10
parameterized==0.8.1
phonenumbers==8.12.50
pillow==6.2.1
psycopg2==2.9.3
pycountry==22.3.5
python-cron==1.0.1
python-dateutil==2.8.2
python-slugify==6.1.2
stripe==6.4.0
twilio==7.10.0
tzdata==2025.2